Avoid checking field manager if field cannot exert force
If a particle is not susceptible to electromagnetic or gravitational
forces (e.g. photons, neutrinos, etc), then we can avoid checking the
field manager altogether, even when a field is present.

Also, the variables restMass, magneticMoment, particleCharge are still
in the cache at this point, while the members may need to be fetched
from elsewhere. Therefore, we should access members like fUseGravity
only after we checked that the particle actually has mass, etc.