Move zero length step case to the top
If currentMinimumStep is zero, it doesn't matter if there is a field or
not, we can just immediately take the zero step and save on unnecessary
checks that were being done before. Note that currentMinimumStep is
passed by value into ComputeStep functions, so it cannot be modified by
those function calls. Consequently, we can also remove the if statement
that flags if the step is limited by geometry towards the end of the
method.