Hoist updates to fEndGlobalTimeComputed and fTransportEndKineticEnergy out of if statement
This allows to ifdef out the else clause of the statement, since it now
only checks for the difference in energy and reports when verbosity is
enabled. The else clause can be enabled back by either setting G4VERBOSE
or G4DEBUG_TRANSPORT if needed. Since this code is performance sensitive,
it makes sense to disable if it's not needed, especially because there
are two static thread local variables that can be eliminated this way,
one of which is not in use (no_inexact_step).